+++ replicante/src/components/mod.rs
@@ -9,6 +9,7 @@ use replicante_data_fetcher::Fetcher;
 use super::Config;
 use super::Interfaces;
 use super::Result;
+use super::metrics::COMPONENTS_ENABLED;
 
 
 pub mod discovery;
@@ -16,6 +17,44 @@ pub mod webui;
 
 use self::discovery::DiscoveryComponent as Discovery;
 use self::webui::WebUI;
+
+
+/// Helper macro to keep `Components::run` simpler in the presence of optional components.
+macro_rules! component_run {
+    ($component:expr) => {
+        if let Some(component) = $component {
+            component.run()?;
+        }
+    };
+}
+
+/// Helper macro to keep `Components::wait_all` simpler in the presence of optional components.
+macro_rules! component_wait {
+    ($component:expr) => {
+        if let Some(component) = $component {
+            component.wait()?;
+        }
+    };
+}
+
+/// Helper function to keep `Components::new` simpler in the presence of optional components.
+fn component_new<C, F>(
+    component: &str, mode: &str, enabled: bool, logger: Logger, factory: F
+) -> Option<C>
+    where F: FnOnce() -> C,
+{
+    info!(
+        logger, "Initialising component if enabled";
+        "component" => component, "type" => mode, "enabled" => enabled
+    );
+    if enabled {
+        COMPONENTS_ENABLED.with_label_values(&[component, mode]).set(1.0);
+        Some(factory())
+    } else {
+        COMPONENTS_ENABLED.with_label_values(&[component, mode]).set(0.0);
+        None
+    }
+}
 
 
 /// A container for replicante components.
@@ -28,18 +67,25 @@ use self::webui::WebUI;
 /// [`Drop`]: std/ops/trait.Drop.html
 /// [`JoinHandle`]: std/thread/struct.JoinHandle.html
 pub struct Components {
-    discovery: Discovery,
-    webui: WebUI,
+    discovery: Option<Discovery>,
+    webui: Option<WebUI>,
 }
 
 impl Components {
     /// Creates and configures components.
     pub fn new(config: &Config, logger: Logger, interfaces: &mut Interfaces) -> Result<Components> {
-        let discovery = Discovery::new(
-            config.discovery.clone(), config.events.snapshots.clone(),
-            Duration::from_secs(config.timeouts.agents_api), logger, interfaces
+        let discovery = component_new(
+            "discovery", "required", config.components.discovery(), logger.clone(), || {
+                Discovery::new(
+                    config.discovery.clone(), config.events.snapshots.clone(),
+                    Duration::from_secs(config.timeouts.agents_api), logger.clone(), interfaces
+                )
+            }
         );
-        let webui = WebUI::new(interfaces);
+        let webui = component_new(
+            "webui", "optional", config.components.webui(), logger.clone(),
+            || WebUI::new(interfaces)
+        );
         Ok(Components {
             discovery,
             webui,
@@ -59,15 +105,15 @@ impl Components {
 
     /// Performs any final configuration and starts background threads.
     pub fn run(&mut self) -> Result<()> {
-        self.discovery.run()?;
-        self.webui.run()?;
+        component_run!(self.discovery.as_mut());
+        component_run!(self.webui.as_mut());
         Ok(())
     }
 
     /// Waits for all interfaces to terminate.
     pub fn wait_all(&mut self) -> Result<()> {
-        self.discovery.wait()?;
-        self.webui.wait()?;
+        component_wait!(self.discovery.as_mut());
+        component_wait!(self.webui.as_mut());
         Ok(())
     }
 }